backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
supabase==2.3.4
python-dotenv==1.0.0
lxml==4.9.3